JAM Testnet - JAMNP Test Vectors DRAFT
This describes some DRAFT **JAM Testnet** test vector files, representing the majority of JAMNP request-response patterns as described in JAM Simple Networking Protocol (JAMNP-S).

It is intended to develop a course of action for which **JAMNP** request-response patterns and erasure coding vectors are necessary to build a multi-team JAM tiny testnet #69.