erasure coding simple check against davxy+javajam (#129)
Attempt to achieve encoding alignment with davxy and javajam EC vectors

* [test_davxy](https://github.com/davxy/jam-test-vectors/tree/erasure_coding) -- specifically test_segment_shards_tiny.json
* [test_javajam](https://github.com/javajamio/javajam-trace/tree/main/erasure_coding) -- specifically segment_tiny.json

Status:
* neither encoding process results in the recovery shards 

See README

Probably there is a conceptual model, a package config difference, between 3 of us.  


While GP Appendix H aims to be a full specification of the EC used, in the end implementers just will FFI into this package rather than implement it in accordance with Appendix H as per rule 1.  So need to figure out the conceptual or model error and socialize the solution.  In this case, its acceptable to publish a code solution that matches some public encoding of segment and work package.

The segment 4104 case is the easiest case to get on the same page about.