0.6.6 31-byte state key dataset (#211)
davxy chose a genesis format with 00000000.json and a statetransition of 31-byte statekeys without metadata, which motivates this refactor:
* `state_transitions` (JSON+bin) be EXACTLY like davxy, but now *without any metadata and full use of 31-bytes
* `state_snapshots` be purely JSON about
  (A) C1-C15 
  (B) state updates:
     - service info 
     - values
     - preimage additions/removal (timeslots)
It is not intended that the state_snapshots can derive the stateroot anymore.  However, the state_transitions (either format) can.

The genesis block/state is in 00000000.json and eliminates the need to publish chainspecs.  

Note: This is a work in progress and an intermediate step towards 0.6.6.